requirements.txt
boto3==1.5.20
nltk==3.2.4
numpy==1.13.1
gensim==2.2.0
pytorch==0.2.0
torchtext==0.2.1
